food-monitor/api/requirements.txt
flask
mysql-connector-python
flask-swagger-ui
